requirements.txt
beautifulsoup4==4.11.1
chromedriver_autoinstaller==0.6.2
discord.py==2.3.2
langchain==0.0.171
pandas==1.4.2
python_json_logger==2.0.7
selenium==4.12.0
